Replaced "Follow up requested date" with "Follow up required: Y".
Removed the date because according to the specification:
"Don't need a 'follow up call request date' as follow up doesn't necessarily refer to a call"

Also made the "Follow up required: Y" show up only when follow-up is actually needed, rather than show "Follow up required: N" when it's not needed, because that's the user wanted. I assume it has to do with decluttering the UI so it would be easiert to find information.