Changed "Is follow-up required check" to be model.CallbackRequired value.
Due to form related changes (parallel & separate PR), it was agreed that the "Follow Required" state should be stored on the database inside already existing model.CallbackRequired field.

As a result the subsequent checks should reflect that by checking for follow by looking into that field from the DB rather than doing a check on whether any of the calls made under the case contain call resolution of "Follow-up required". (which was flawed to begin with due to it checking accross all calls in the history instead of the latest one).